Refactor NonCommands
Now every NonCommand class is a Spring Component that store a set of ChatStates that it handles. This allows to autowire these classes into Set as a NonCommandUpdateHandler's constructor's parameter.

As a result: now everything I need to add a new NonCommand is to develop a new class extending AbstractNonCommand. I don't have to add info to Config class etc.